brought backend call to app.jsx for better accesability throughout entire app
